Post office hours
Stackoverflow suggsted changing the data toggle in the dropdown from data-toggle to data-bs-toggle. This fixed the issue of the dropdown not working. However, the hamburger menu does not work. Uncommenting the links above and the scripts below, and vise versa, fixes the hamburger meu. However, this will break the dropdown unless the -bs- change is reverted. Still looking for more answers to fix everything while still using the local host and most recent versions of bootstrap and everything else.